W-12277411: Expose the data about possible drops of spans (#1980)
(cherry picked from commit 6dd80ddaf2fbf57e39cde36eccbaf67ba276560f)
