+++ tracing/src/test/java/org/mule/test/components/tracing/AsyncErrorOpenTelemetryTracingTestCase.java
@@ -12,7 +12,9 @@ import static org.mule.test.allure.AllureConstants.Profiling.ProfilingServiceSto
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.Assert.assertThat;
 
+import org.mule.functional.api.flow.FlowRunner;
 import org.mule.functional.junit4.MuleArtifactFunctionalTestCase;
+import org.mule.runtime.core.privileged.event.BaseEventContext;
 import org.mule.runtime.tracer.api.sniffer.CapturedExportedSpan;
 import org.mule.runtime.tracer.api.sniffer.ExportedSpanSniffer;
 import org.mule.runtime.core.privileged.profiling.PrivilegedProfilingService;
@@ -22,6 +24,7 @@ import org.mule.tck.probe.PollingProber;
 import org.mule.test.infrastructure.profiling.tracing.SpanTestHierarchy;
 
 import java.util.Collection;
+import java.util.concurrent.CountDownLatch;
 
 import javax.inject.Inject;
 
@@ -31,36 +34,41 @@ import org.junit.Test;
 
 @Feature(PROFILING)
 @Story(DEFAULT_CORE_EVENT_TRACER)
-public class ScatterGatherDoubleErrorTracingTestCase extends MuleArtifactFunctionalTestCase
-    implements TracingTestRunnerConfigAnnotation {
+public class AsyncErrorOpenTelemetryTracingTestCase extends MuleArtifactFunctionalTestCase
+    implements OpenTelemetryTracingTestRunnerConfigAnnotation {
 
   private static final int TIMEOUT_MILLIS = 30000;
   private static final int POLL_DELAY_MILLIS = 100;
 
-  public static final String EXPECTED_ROUTE_SPAN_NAME = "mule:scatter-gather:route";
-  public static final String EXPECTED_SCATTER_GATHER_SPAN_NAME = "mule:scatter-gather";
+
+  public static final String EXPECTED_ASYNC_SPAN_NAME = "mule:async";
   public static final String EXPECTED_LOGGER_SPAN_NAME = "mule:logger";
   public static final String EXPECTED_SET_PAYLOAD_SPAN_NAME = "mule:set-payload";
-  public static final String SCATTER_GATHER_FLOW = "scatter-gather-flow";
+  public static final String EXPECTED_SET_VARIABLE_SPAN_NAME = "mule:set-variable";
+  public static final String ASYNC_FLOW = "async-flow";
   public static final String EXPECTED_FLOW_SPAN_NAME = "mule:flow";
-  public static final String EXPECTED_ON_ERROR_PROPAGATE_SPAN_NAME = "mule:on-error-propagate";
   public static final String NO_PARENT_SPAN = "0000000000000000";
-  public static final String EXPECTED_RAISE_ERROR_SPAN = "mule:raise-error";
+  private static final String EXPECTED_RAISE_ERROR_SPAN_NAME = "mule:raise-error";
 
   @Inject
   PrivilegedProfilingService profilingService;
 
   @Override
   protected String getConfigFile() {
-    return "tracing/scatter-gather-double-error.xml";
+    return "tracing/async-error.xml";
   }
 
   @Test
-  public void testScatterGatherFlowWithError() throws Exception {
+  public void testAsyncFlow() throws Exception {
     ExportedSpanSniffer spanCapturer = profilingService.getSpanExportManager().getExportedSpanSniffer();
 
     try {
-      flowRunner(SCATTER_GATHER_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD).dispatch();
+      CountDownLatch asyncTerminationLatch = new CountDownLatch(1);
+      FlowRunner runner = flowRunner(ASYNC_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD);
+      ((BaseEventContext) (runner.buildEvent().getContext())).onTerminated((e, t) -> asyncTerminationLatch.countDown());
+      runner.run();
+
+      asyncTerminationLatch.await();
 
       PollingProber prober = new PollingProber(TIMEOUT_MILLIS, POLL_DELAY_MILLIS);
 
@@ -69,7 +77,7 @@ public class ScatterGatherDoubleErrorTracingTestCase extends MuleArtifactFunctio
         @Override
         protected boolean test() {
           Collection<CapturedExportedSpan> exportedSpans = spanCapturer.getExportedSpans();;
-          return exportedSpans.size() == 8;
+          return exportedSpans.size() == 6;
         }
 
         @Override
@@ -78,24 +86,19 @@ public class ScatterGatherDoubleErrorTracingTestCase extends MuleArtifactFunctio
         }
       });
 
+
       Collection<CapturedExportedSpan> exportedSpans = spanCapturer.getExportedSpans();
 
       SpanTestHierarchy expectedSpanHierarchy = new SpanTestHierarchy(exportedSpans);
       expectedSpanHierarchy.withRoot(EXPECTED_FLOW_SPAN_NAME)
           .beginChildren()
-          .child(EXPECTED_SCATTER_GATHER_SPAN_NAME)
+          .child(EXPECTED_SET_VARIABLE_SPAN_NAME)
+          .child(EXPECTED_ASYNC_SPAN_NAME)
           .beginChildren()
-          .child(EXPECTED_ROUTE_SPAN_NAME)
-          .beginChildren()
-          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
-          .child(EXPECTED_RAISE_ERROR_SPAN)
+          .child(EXPECTED_LOGGER_SPAN_NAME)
+          .child(EXPECTED_RAISE_ERROR_SPAN_NAME)
           .endChildren()
-          .child(EXPECTED_ROUTE_SPAN_NAME)
-          .beginChildren()
-          .child(EXPECTED_RAISE_ERROR_SPAN)
-          .endChildren()
-          .endChildren()
-          .child(EXPECTED_ON_ERROR_PROPAGATE_SPAN_NAME)
+          .child(EXPECTED_SET_VARIABLE_SPAN_NAME)
           .endChildren();
 
       expectedSpanHierarchy.assertSpanTree();
