+++ tracing/src/test/java/org/mule/test/components/tracing/AsyncSuccessOpenTelemetryTracingTestCase.java
@@ -4,6 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
+
 package org.mule.test.components.tracing;
 
 import static org.mule.test.allure.AllureConstants.Profiling.PROFILING;
@@ -12,7 +13,9 @@ import static org.mule.test.allure.AllureConstants.Profiling.ProfilingServiceSto
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.Assert.assertThat;
 
+import org.mule.functional.api.flow.FlowRunner;
 import org.mule.functional.junit4.MuleArtifactFunctionalTestCase;
+import org.mule.runtime.core.privileged.event.BaseEventContext;
 import org.mule.runtime.tracer.api.sniffer.CapturedExportedSpan;
 import org.mule.runtime.tracer.api.sniffer.ExportedSpanSniffer;
 import org.mule.runtime.core.privileged.profiling.PrivilegedProfilingService;
@@ -22,6 +25,7 @@ import org.mule.tck.probe.PollingProber;
 import org.mule.test.infrastructure.profiling.tracing.SpanTestHierarchy;
 
 import java.util.Collection;
+import java.util.concurrent.CountDownLatch;
 
 import javax.inject.Inject;
 
@@ -31,34 +35,38 @@ import org.junit.Test;
 
 @Feature(PROFILING)
 @Story(DEFAULT_CORE_EVENT_TRACER)
-public class ScatterGatherSuccessTracingTestCase extends MuleArtifactFunctionalTestCase
-    implements TracingTestRunnerConfigAnnotation {
+public class AsyncSuccessOpenTelemetryTracingTestCase extends MuleArtifactFunctionalTestCase
+    implements OpenTelemetryTracingTestRunnerConfigAnnotation {
 
   private static final int TIMEOUT_MILLIS = 30000;
   private static final int POLL_DELAY_MILLIS = 100;
 
-  public static final String EXPECTED_ROUTE_SPAN_NAME = "mule:scatter-gather:route";
-  public static final String EXPECTED_SCATTER_GATHER_SPAN_NAME = "mule:scatter-gather";
+  public static final String EXPECTED_ASYNC_SPAN_NAME = "mule:async";
   public static final String EXPECTED_LOGGER_SPAN_NAME = "mule:logger";
   public static final String EXPECTED_SET_PAYLOAD_SPAN_NAME = "mule:set-payload";
-  public static final String SCATTER_GATHER_FLOW = "scatter-gather-flow";
+  public static final String EXPECTED_SET_VARIABLE_SPAN_NAME = "mule:set-variable";
+  public static final String ASYNC_FLOW = "async-flow";
   public static final String EXPECTED_FLOW_SPAN_NAME = "mule:flow";
-  public static final String NO_PARENT_SPAN = "0000000000000000";
 
   @Inject
   PrivilegedProfilingService profilingService;
 
   @Override
   protected String getConfigFile() {
-    return "tracing/scatter-gather-success.xml";
+    return "tracing/async-success.xml";
   }
 
   @Test
-  public void testScatterGatherFlow() throws Exception {
+  public void testAsyncFlow() throws Exception {
     ExportedSpanSniffer spanCapturer = profilingService.getSpanExportManager().getExportedSpanSniffer();
 
     try {
-      flowRunner(SCATTER_GATHER_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD).run().getMessage();
+      CountDownLatch asyncTerminationLatch = new CountDownLatch(1);
+      FlowRunner runner = flowRunner(ASYNC_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD);
+      ((BaseEventContext) (runner.buildEvent().getContext())).onTerminated((e, t) -> asyncTerminationLatch.countDown());
+      runner.run();
+
+      asyncTerminationLatch.await();
 
       PollingProber prober = new PollingProber(TIMEOUT_MILLIS, POLL_DELAY_MILLIS);
 
@@ -67,7 +75,7 @@ public class ScatterGatherSuccessTracingTestCase extends MuleArtifactFunctionalT
         @Override
         protected boolean test() {
           Collection<CapturedExportedSpan> exportedSpans = spanCapturer.getExportedSpans();;
-          return exportedSpans.size() == 7;
+          return exportedSpans.size() == 6;
         }
 
         @Override
@@ -81,18 +89,12 @@ public class ScatterGatherSuccessTracingTestCase extends MuleArtifactFunctionalT
       SpanTestHierarchy expectedSpanHierarchy = new SpanTestHierarchy(exportedSpans);
       expectedSpanHierarchy.withRoot(EXPECTED_FLOW_SPAN_NAME)
           .beginChildren()
-          .child(EXPECTED_SCATTER_GATHER_SPAN_NAME)
+          .child(EXPECTED_SET_VARIABLE_SPAN_NAME)
+          .child(EXPECTED_SET_VARIABLE_SPAN_NAME)
+          .child(EXPECTED_ASYNC_SPAN_NAME)
           .beginChildren()
-          .child(EXPECTED_ROUTE_SPAN_NAME)
-          .beginChildren()
-          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
-          .endChildren()
-          .child(EXPECTED_ROUTE_SPAN_NAME)
-          .beginChildren()
-          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
-          .endChildren()
-          .endChildren()
           .child(EXPECTED_LOGGER_SPAN_NAME)
+          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
           .endChildren();
 
       expectedSpanHierarchy.assertSpanTree();
