+++ tracing/src/test/java/org/mule/test/components/tracing/DistributedOpenTelemetryTracingTestCase.java
@@ -12,8 +12,6 @@ import static org.mule.runtime.tracer.exporter.api.config.OpenTelemetrySpanExpor
 import static org.mule.runtime.tracer.exporter.api.config.OpenTelemetrySpanExporterConfigurationProperties.MULE_OPEN_TELEMETRY_EXPORTER_ENDPOINT;
 import static org.mule.runtime.tracer.exporter.api.config.OpenTelemetrySpanExporterConfigurationProperties.MULE_OPEN_TELEMETRY_EXPORTER_TYPE;
 import static org.mule.test.components.tracing.OpenTelemetryProtobufSpanUtils.getSpans;
-import static org.mule.test.infrastructure.profiling.tracing.TracingTestUtils.createAttributeMap;
-import static org.mule.test.infrastructure.profiling.tracing.TracingTestUtils.getDefaultAttributesToAssertExistence;
 
 import static java.lang.Boolean.TRUE;
 import static java.lang.String.valueOf;
@@ -23,11 +21,10 @@ import static java.util.Arrays.asList;
 
 import static com.linecorp.armeria.common.HttpResponse.from;
 import static com.linecorp.armeria.common.HttpStatus.OK;
-import static com.linecorp.armeria.common.HttpStatus.REQUEST_TIMEOUT;
 
-import org.junit.After;
 import org.mule.functional.junit4.MuleArtifactFunctionalTestCase;
 import org.mule.runtime.tracer.api.sniffer.CapturedExportedSpan;
+import org.mule.tck.junit4.rule.DynamicPort;
 import org.mule.tck.probe.JUnitProbe;
 import org.mule.tck.probe.PollingProber;
 import org.mule.test.infrastructure.profiling.tracing.SpanTestHierarchy;
@@ -38,9 +35,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import com.linecorp.armeria.common.HttpRequest;
 import com.linecorp.armeria.common.HttpResponse;
@@ -50,33 +45,38 @@ import com.linecorp.armeria.server.ServiceRequestContext;
 import com.linecorp.armeria.testing.junit4.server.ServerRule;
 import io.opentelemetry.proto.collector.trace.v1.ExportTraceServiceRequest;
 import org.jetbrains.annotations.NotNull;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.ClassRule;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runners.Parameterized;
 
 @RunnerDelegateTo(Parameterized.class)
-public class RetryTracingTestCase extends
-    MuleArtifactFunctionalTestCase implements TracingTestRunnerConfigAnnotation {
+public class DistributedOpenTelemetryTracingTestCase extends
+    MuleArtifactFunctionalTestCase implements OpenTelemetryTracingTestRunnerConfigAnnotation {
 
-  private static final String EXPECTED_FLOW_SPAN_NAME = "mule:flow";
-  private static final String EXPECTED_SET_PAYLOAD_SPAN_NAME = "mule:set-payload";
-  private static final String FLOW_LOCATION = "retry-flow";
+  @Rule
+  public DynamicPort httpPort = new DynamicPort("port");
 
-  private static final String TEST_ARTIFACT_ID = "RetryTracingTestCase#testRetryBackoffTest";
+  private static final String STARTING_FLOW = "startingFlow";
+
+  private static final String EXPECTED_HTTP_REQUEST_SPAN_NAME = "HTTP GET";
+  private static final String EXPECTED_HTTP_FLOW_SPAN_NAME = "/test";
+  private static final String EXPECTED_LOGGER_SPAN_NAME = "mule:logger";
+  public static final String EXPECTED_FLOW_SPAN_NAME = "mule:flow";
 
   public static final int TIMEOUT_MILLIS = 30000;
 
   private static final int POLL_DELAY_MILLIS = 100;
   public static final int MAX_BACKOFF_ATTEMPTS = 2;
 
-  private static final String SET_PAYLOAD_LOCATION = "retry-flow/processors/0";
   private final String type;
   private final String path;
 
   @Override
   protected String getConfigFile() {
-    return "tracing/retry-backoff.xml";
+    return "tracing/distributed-tracing.xml";
   }
 
   @ClassRule
@@ -90,7 +90,7 @@ public class RetryTracingTestCase extends
     });
   }
 
-  public RetryTracingTestCase(String type, String path) {
+  public DistributedOpenTelemetryTracingTestCase(String type, String path) {
     this.type = type;
     this.path = path;
   }
@@ -113,8 +113,8 @@ public class RetryTracingTestCase extends
   }
 
   @Test
-  public void testRetryBackoffTest() throws Exception {
-    flowRunner(FLOW_LOCATION).withPayload(TEST_PAYLOAD).run().getMessage();
+  public void testDistributedTracing() throws Exception {
+    flowRunner(STARTING_FLOW).run();
     PollingProber prober = new PollingProber(TIMEOUT_MILLIS, POLL_DELAY_MILLIS);
 
     prober.check(new JUnitProbe() {
@@ -122,7 +122,7 @@ public class RetryTracingTestCase extends
       @Override
       protected boolean test() {
         Collection<CapturedExportedSpan> exportedSpans = httpServer.getCapturedExportedSpans();
-        return exportedSpans.size() == 2;
+        return exportedSpans.size() == 4;
       }
 
       @Override
@@ -131,22 +131,18 @@ public class RetryTracingTestCase extends
       }
     });
 
-    List<String> attributesToAssertExistence = getDefaultAttributesToAssertExistence();
-
     Collection<CapturedExportedSpan> exportedSpans = httpServer.getCapturedExportedSpans();
 
-    String artifactId = TEST_ARTIFACT_ID + "[" + type + "]";
-
-    Map<String, String> setPayloadAttributeMap = createAttributeMap(SET_PAYLOAD_LOCATION, artifactId);
 
     SpanTestHierarchy expectedSpanHierarchy = new SpanTestHierarchy(exportedSpans);
     expectedSpanHierarchy.withRoot(EXPECTED_FLOW_SPAN_NAME)
-        .addAttributesToAssertValue(createAttributeMap(FLOW_LOCATION, artifactId))
-        .addAttributesToAssertExistence(attributesToAssertExistence)
         .beginChildren()
-        .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
-        .addAttributesToAssertValue(setPayloadAttributeMap)
-        .addAttributesToAssertExistence(attributesToAssertExistence)
+        .child(EXPECTED_HTTP_REQUEST_SPAN_NAME)
+        .beginChildren()
+        .child(EXPECTED_HTTP_FLOW_SPAN_NAME)
+        .beginChildren()
+        .child(EXPECTED_LOGGER_SPAN_NAME)
+        .endChildren()
         .endChildren();
 
     expectedSpanHierarchy.assertSpanTree();
@@ -158,8 +154,6 @@ public class RetryTracingTestCase extends
 
     private final List<CapturedExportedSpan> capturedExportedSpans = new ArrayList<>();
 
-    private final AtomicInteger exportAttempts = new AtomicInteger(0);
-
     @Override
     protected void configure(ServerBuilder sb) {
       sb.service(PATH_PATTERN,
@@ -170,10 +164,6 @@ public class RetryTracingTestCase extends
                      return HttpResponse.from(req.aggregate().handle((aReq, cause) -> {
                        CompletableFuture<HttpResponse> responseFuture = new CompletableFuture<>();
                        HttpResponse res = from(responseFuture);
-                       if (exportAttempts.incrementAndGet() < MAX_BACKOFF_ATTEMPTS) {
-                         responseFuture.complete(HttpResponse.of(REQUEST_TIMEOUT));
-                         return res;
-                       }
                        try {
                          capturedExportedSpans.addAll(getSpans(ExportTraceServiceRequest
                              .parseFrom(new ByteArrayInputStream(aReq.content().array()))));
@@ -182,7 +172,6 @@ public class RetryTracingTestCase extends
                        }
                        responseFuture.complete(HttpResponse.of(OK));
                        return res;
-
                      }));
                    }
                  });
