+++ tracing/src/test/java/org/mule/test/components/tracing/FirstSuccessfulDoubleErrorOpenTelemetryTracingTestCase.java
@@ -4,6 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
+
 package org.mule.test.components.tracing;
 
 import static org.mule.test.allure.AllureConstants.Profiling.PROFILING;
@@ -12,9 +13,7 @@ import static org.mule.test.allure.AllureConstants.Profiling.ProfilingServiceSto
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.Assert.assertThat;
 
-import org.mule.functional.api.flow.FlowRunner;
 import org.mule.functional.junit4.MuleArtifactFunctionalTestCase;
-import org.mule.runtime.core.privileged.event.BaseEventContext;
 import org.mule.runtime.tracer.api.sniffer.CapturedExportedSpan;
 import org.mule.runtime.tracer.api.sniffer.ExportedSpanSniffer;
 import org.mule.runtime.core.privileged.profiling.PrivilegedProfilingService;
@@ -24,7 +23,6 @@ import org.mule.tck.probe.PollingProber;
 import org.mule.test.infrastructure.profiling.tracing.SpanTestHierarchy;
 
 import java.util.Collection;
-import java.util.concurrent.CountDownLatch;
 
 import javax.inject.Inject;
 
@@ -34,40 +32,39 @@ import org.junit.Test;
 
 @Feature(PROFILING)
 @Story(DEFAULT_CORE_EVENT_TRACER)
-public class AsyncErrorTracingTestCase extends MuleArtifactFunctionalTestCase implements TracingTestRunnerConfigAnnotation {
+public class FirstSuccessfulDoubleErrorOpenTelemetryTracingTestCase extends MuleArtifactFunctionalTestCase
+    implements OpenTelemetryTracingTestRunnerConfigAnnotation {
 
   private static final int TIMEOUT_MILLIS = 30000;
   private static final int POLL_DELAY_MILLIS = 100;
 
-
-  public static final String EXPECTED_ASYNC_SPAN_NAME = "mule:async";
+  public static final String EXPECTED_ROUTE_SPAN_NAME_ATTEMPT_1 = "mule:first-successful:attempt:1";
+  public static final String EXPECTED_ROUTE_SPAN_NAME_ATTEMPT_2 = "mule:first-successful:attempt:2";
+  public static final String EXPECTED_FIRST_SUCCESSFUL_SPAN_NAME = "mule:first-successful";
   public static final String EXPECTED_LOGGER_SPAN_NAME = "mule:logger";
-  public static final String EXPECTED_SET_PAYLOAD_SPAN_NAME = "mule:set-payload";
-  public static final String EXPECTED_SET_VARIABLE_SPAN_NAME = "mule:set-variable";
-  public static final String ASYNC_FLOW = "async-flow";
+  public static final String FLOW = "first-successful-telemetryFlow";
   public static final String EXPECTED_FLOW_SPAN_NAME = "mule:flow";
+  public static final String EXPECTED_SET_VARIABLE_SPAN_NAME = "mule:set-variable";
+  public static final String EXPECTED_SET_PAYLOAD_SPAN_NAME = "mule:set-payload";
+  public static final String EXPECTED_ON_ERROR_PROPAGATE_SPAN_NAME = "mule:on-error-propagate";
   public static final String NO_PARENT_SPAN = "0000000000000000";
-  private static final String EXPECTED_RAISE_ERROR_SPAN_NAME = "mule:raise-error";
+  public static final int NUMBER_OF_ROUTES = 2;
+  public static final String EXPECTED_RAISE_ERROR_SPAN = "mule:raise-error";
 
   @Inject
   PrivilegedProfilingService profilingService;
 
   @Override
   protected String getConfigFile() {
-    return "tracing/async-error.xml";
+    return "tracing/first-successful-double-error.xml";
   }
 
   @Test
-  public void testAsyncFlow() throws Exception {
+  public void testFlow() throws Exception {
     ExportedSpanSniffer spanCapturer = profilingService.getSpanExportManager().getExportedSpanSniffer();
 
     try {
-      CountDownLatch asyncTerminationLatch = new CountDownLatch(1);
-      FlowRunner runner = flowRunner(ASYNC_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD);
-      ((BaseEventContext) (runner.buildEvent().getContext())).onTerminated((e, t) -> asyncTerminationLatch.countDown());
-      runner.run();
-
-      asyncTerminationLatch.await();
+      flowRunner(FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD).dispatch();
 
       PollingProber prober = new PollingProber(TIMEOUT_MILLIS, POLL_DELAY_MILLIS);
 
@@ -76,7 +73,7 @@ public class AsyncErrorTracingTestCase extends MuleArtifactFunctionalTestCase im
         @Override
         protected boolean test() {
           Collection<CapturedExportedSpan> exportedSpans = spanCapturer.getExportedSpans();;
-          return exportedSpans.size() == 6;
+          return exportedSpans.size() == 3 * NUMBER_OF_ROUTES + 5;
         }
 
         @Override
@@ -85,19 +82,27 @@ public class AsyncErrorTracingTestCase extends MuleArtifactFunctionalTestCase im
         }
       });
 
-
       Collection<CapturedExportedSpan> exportedSpans = spanCapturer.getExportedSpans();
 
       SpanTestHierarchy expectedSpanHierarchy = new SpanTestHierarchy(exportedSpans);
       expectedSpanHierarchy.withRoot(EXPECTED_FLOW_SPAN_NAME)
           .beginChildren()
           .child(EXPECTED_SET_VARIABLE_SPAN_NAME)
-          .child(EXPECTED_ASYNC_SPAN_NAME)
+          .child(EXPECTED_FIRST_SUCCESSFUL_SPAN_NAME)
           .beginChildren()
-          .child(EXPECTED_LOGGER_SPAN_NAME)
-          .child(EXPECTED_RAISE_ERROR_SPAN_NAME)
+          .child(EXPECTED_ROUTE_SPAN_NAME_ATTEMPT_1)
+          .beginChildren()
+          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
+          .child(EXPECTED_RAISE_ERROR_SPAN)
           .endChildren()
-          .child(EXPECTED_SET_VARIABLE_SPAN_NAME)
+          .child(EXPECTED_ROUTE_SPAN_NAME_ATTEMPT_2)
+          .beginChildren()
+          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
+          .child(EXPECTED_LOGGER_SPAN_NAME)
+          .child(EXPECTED_RAISE_ERROR_SPAN)
+          .endChildren()
+          .endChildren()
+          .child(EXPECTED_ON_ERROR_PROPAGATE_SPAN_NAME)
           .endChildren();
 
       expectedSpanHierarchy.assertSpanTree();
