+++ tracing/src/test/java/org/mule/test/components/tracing/OpenTelemetryJmsSemanticConventionAttributesAndNameTestCase.java
@@ -41,8 +41,8 @@ import org.junit.Test;
 
 @Feature(PROFILING)
 @Story(DEFAULT_CORE_EVENT_TRACER)
-public class JmsSemanticConventionAttributesAndNameTestCase extends MuleArtifactFunctionalTestCase
-    implements TracingTestRunnerConfigAnnotation {
+public class OpenTelemetryJmsSemanticConventionAttributesAndNameTestCase extends MuleArtifactFunctionalTestCase
+    implements OpenTelemetryTracingTestRunnerConfigAnnotation {
 
   private static final int TIMEOUT_MILLIS = 30000;
   private static final int POLL_DELAY_MILLIS = 100;
