+++ tracing/src/test/java/org/mule/test/components/tracing/ParallelForEachSuccessOpenTelemetryTracingTestCase.java
@@ -4,7 +4,6 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-
 package org.mule.test.components.tracing;
 
 import static org.mule.test.allure.AllureConstants.Profiling.PROFILING;
@@ -32,30 +31,27 @@ import org.junit.Test;
 
 @Feature(PROFILING)
 @Story(DEFAULT_CORE_EVENT_TRACER)
-public class FirstSuccessfulErrorTracingTestCase extends MuleArtifactFunctionalTestCase
-    implements TracingTestRunnerConfigAnnotation {
+public class ParallelForEachSuccessOpenTelemetryTracingTestCase extends MuleArtifactFunctionalTestCase
+    implements OpenTelemetryTracingTestRunnerConfigAnnotation {
 
   private static final int TIMEOUT_MILLIS = 30000;
   private static final int POLL_DELAY_MILLIS = 100;
 
-  public static final String EXPECTED_ROUTE_SPAN_NAME_ATTEMPT_1 = "mule:first-successful:attempt:1";
-  public static final String EXPECTED_ROUTE_SPAN_NAME_ATTEMPT_2 = "mule:first-successful:attempt:2";
-  public static final String EXPECTED_FIRST_SUCCESSFUL_SPAN_NAME = "mule:first-successful";
+  public static final String EXPECTED_ROUTE_SPAN_NAME = "mule:parallel-foreach:iteration";
+  public static final String EXPECTED_PARALLEL_FOREACH_SPAN_NAME = "mule:parallel-foreach";
   public static final String EXPECTED_LOGGER_SPAN_NAME = "mule:logger";
-  public static final String FLOW = "first-successful-telemetryFlow";
+  public static final String PARALLEL_FOR_EACH_FLOW = "parallel-for-eachFlow";
   public static final String EXPECTED_FLOW_SPAN_NAME = "mule:flow";
-  public static final String EXPECTED_SET_VARIABLE_SPAN_NAME = "mule:set-variable";
   public static final String EXPECTED_SET_PAYLOAD_SPAN_NAME = "mule:set-payload";
   public static final String NO_PARENT_SPAN = "0000000000000000";
-  public static final int NUMBER_OF_ROUTES = 2;
-  public static final String EXPECTED_RAISE_ERROR_SPAN = "mule:raise-error";
+  public static final int NUMBER_OF_ROUTES = 3;
 
   @Inject
   PrivilegedProfilingService profilingService;
 
   @Override
   protected String getConfigFile() {
-    return "tracing/first-successful-error.xml";
+    return "tracing/parallel-foreach-success.xml";
   }
 
   @Test
@@ -63,7 +59,7 @@ public class FirstSuccessfulErrorTracingTestCase extends MuleArtifactFunctionalT
     ExportedSpanSniffer spanCapturer = profilingService.getSpanExportManager().getExportedSpanSniffer();
 
     try {
-      flowRunner(FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD).dispatch();
+      flowRunner(PARALLEL_FOR_EACH_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD).dispatch();
 
       PollingProber prober = new PollingProber(TIMEOUT_MILLIS, POLL_DELAY_MILLIS);
 
@@ -72,7 +68,7 @@ public class FirstSuccessfulErrorTracingTestCase extends MuleArtifactFunctionalT
         @Override
         protected boolean test() {
           Collection<CapturedExportedSpan> exportedSpans = spanCapturer.getExportedSpans();;
-          return exportedSpans.size() == 3 * NUMBER_OF_ROUTES + 4;
+          return exportedSpans.size() == NUMBER_OF_ROUTES * 3 + 4;
         }
 
         @Override
@@ -86,21 +82,26 @@ public class FirstSuccessfulErrorTracingTestCase extends MuleArtifactFunctionalT
       SpanTestHierarchy expectedSpanHierarchy = new SpanTestHierarchy(exportedSpans);
       expectedSpanHierarchy.withRoot(EXPECTED_FLOW_SPAN_NAME)
           .beginChildren()
-          .child(EXPECTED_SET_VARIABLE_SPAN_NAME)
-          .child(EXPECTED_FIRST_SUCCESSFUL_SPAN_NAME)
+          .child(EXPECTED_PARALLEL_FOREACH_SPAN_NAME)
           .beginChildren()
-          .child(EXPECTED_ROUTE_SPAN_NAME_ATTEMPT_1)
+          .child(EXPECTED_ROUTE_SPAN_NAME)
           .beginChildren()
-          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
-          .child(EXPECTED_RAISE_ERROR_SPAN)
-          .endChildren()
-          .child(EXPECTED_ROUTE_SPAN_NAME_ATTEMPT_2)
-          .beginChildren()
-          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
           .child(EXPECTED_LOGGER_SPAN_NAME)
+          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
+          .endChildren()
+          .child(EXPECTED_ROUTE_SPAN_NAME)
+          .beginChildren()
+          .child(EXPECTED_LOGGER_SPAN_NAME)
+          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
+          .endChildren()
+          .child(EXPECTED_ROUTE_SPAN_NAME)
+          .beginChildren()
+          .child(EXPECTED_LOGGER_SPAN_NAME)
+          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
           .endChildren()
           .endChildren()
-          .child(EXPECTED_SET_VARIABLE_SPAN_NAME)
+          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
+          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
           .endChildren();
 
       expectedSpanHierarchy.assertSpanTree();
