+++ tracing/src/test/java/org/mule/test/components/tracing/RoundRobinErrorOpenTelemetryTracingTestCase.java
@@ -4,12 +4,15 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-
 package org.mule.test.components.tracing;
 
 import static org.mule.test.allure.AllureConstants.Profiling.PROFILING;
 import static org.mule.test.allure.AllureConstants.Profiling.ProfilingServiceStory.DEFAULT_CORE_EVENT_TRACER;
+import static org.mule.test.infrastructure.profiling.tracing.TracingTestUtils.ARTIFACT_ID_KEY;
+import static org.mule.test.infrastructure.profiling.tracing.TracingTestUtils.createAttributeMap;
+import static org.mule.test.infrastructure.profiling.tracing.TracingTestUtils.getDefaultAttributesToAssertExistence;
 
+import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.Assert.assertThat;
 
@@ -23,6 +26,7 @@ import org.mule.tck.probe.PollingProber;
 import org.mule.test.infrastructure.profiling.tracing.SpanTestHierarchy;
 
 import java.util.Collection;
+import java.util.List;
 
 import javax.inject.Inject;
 
@@ -32,37 +36,37 @@ import org.junit.Test;
 
 @Feature(PROFILING)
 @Story(DEFAULT_CORE_EVENT_TRACER)
-public class ParallelForEachErrorTracingTestCase extends MuleArtifactFunctionalTestCase
-    implements TracingTestRunnerConfigAnnotation {
+public class RoundRobinErrorOpenTelemetryTracingTestCase extends MuleArtifactFunctionalTestCase
+    implements OpenTelemetryTracingTestRunnerConfigAnnotation {
 
   private static final int TIMEOUT_MILLIS = 30000;
   private static final int POLL_DELAY_MILLIS = 100;
 
-  public static final String EXPECTED_ROUTE_SPAN_NAME = "mule:parallel-foreach:iteration";
-  public static final String EXPECTED_PARALLEL_FOREACH_SPAN_NAME = "mule:parallel-foreach";
+  public static final String EXPECTED_ROUTE_SPAN_NAME = "mule:round-robin:route";
+  public static final String EXPECTED_ROUND_ROBIN_SPAN_NAME = "mule:round-robin";
   public static final String EXPECTED_LOGGER_SPAN_NAME = "mule:logger";
-  public static final String PARALLEL_FOR_EACH_FLOW = "parallel-for-eachFlow";
+  public static final String ROUND_ROBIN_FLOW = "round-robin-flow";
   public static final String EXPECTED_FLOW_SPAN_NAME = "mule:flow";
-  public static final String EXPECTED_SET_PAYLOAD_SPAN_NAME = "mule:set-payload";
-  public static final String EXPECTED_ON_ERROR_PROPAGATE_SPAN_NAME = "mule:on-error-propagate";
-  public static final String NO_PARENT_SPAN = "0000000000000000";
-  public static final int NUMBER_OF_ROUTES = 3;
   public static final String EXPECTED_RAISE_ERROR_SPAN = "mule:raise-error";
+  public static final String EXPECTED_ON_ERROR_PROPAGATE_SPAN = "mule:on-error-propagate";
+  public static final String NO_PARENT_SPAN = "0000000000000000";
+
+  public static final String TEST_ARTIFACT_ID = "RoundRobinErrorOpenTelemetryTracingTestCase#testRoundRobinFlowWithError";
 
   @Inject
   PrivilegedProfilingService profilingService;
 
   @Override
   protected String getConfigFile() {
-    return "tracing/parallel-foreach-error.xml";
+    return "tracing/round-robin-error.xml";
   }
 
   @Test
-  public void testFlowWithError() throws Exception {
+  public void testRoundRobinFlowWithError() throws Exception {
     ExportedSpanSniffer spanCapturer = profilingService.getSpanExportManager().getExportedSpanSniffer();
 
     try {
-      flowRunner(PARALLEL_FOR_EACH_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD).dispatch();
+      flowRunner(ROUND_ROBIN_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD).runExpectingException();
 
       PollingProber prober = new PollingProber(TIMEOUT_MILLIS, POLL_DELAY_MILLIS);
 
@@ -71,7 +75,7 @@ public class ParallelForEachErrorTracingTestCase extends MuleArtifactFunctionalT
         @Override
         protected boolean test() {
           Collection<CapturedExportedSpan> exportedSpans = spanCapturer.getExportedSpans();;
-          return exportedSpans.size() == NUMBER_OF_ROUTES * 3 + 4;
+          return exportedSpans.size() == 6;
         }
 
         @Override
@@ -82,32 +86,36 @@ public class ParallelForEachErrorTracingTestCase extends MuleArtifactFunctionalT
 
       Collection<CapturedExportedSpan> exportedSpans = spanCapturer.getExportedSpans();
 
+      List<String> attributesToAssertExistence = getDefaultAttributesToAssertExistence();
+
       SpanTestHierarchy expectedSpanHierarchy = new SpanTestHierarchy(exportedSpans);
       expectedSpanHierarchy.withRoot(EXPECTED_FLOW_SPAN_NAME)
+          .addAttributesToAssertValue(createAttributeMap("round-robin-flow", TEST_ARTIFACT_ID))
+          .addAttributesToAssertExistence(attributesToAssertExistence)
           .beginChildren()
-          .child(EXPECTED_SET_PAYLOAD_SPAN_NAME)
-          .child(EXPECTED_PARALLEL_FOREACH_SPAN_NAME)
+          .child(EXPECTED_ROUND_ROBIN_SPAN_NAME)
+          .addAttributesToAssertValue(createAttributeMap("round-robin-flow/processors/0", TEST_ARTIFACT_ID))
+          .addAttributesToAssertExistence(attributesToAssertExistence)
           .beginChildren()
           .child(EXPECTED_ROUTE_SPAN_NAME)
+          .addAttributesToAssertValue(createAttributeMap("round-robin-flow/processors/0", TEST_ARTIFACT_ID))
+          .addAttributesToAssertExistence(attributesToAssertExistence)
           .beginChildren()
           .child(EXPECTED_LOGGER_SPAN_NAME)
+          .addAttributesToAssertValue(createAttributeMap("round-robin-flow/processors/0/route/0/processors/0", TEST_ARTIFACT_ID))
+          .addAttributesToAssertExistence(attributesToAssertExistence)
           .child(EXPECTED_RAISE_ERROR_SPAN)
-          .endChildren()
-          .child(EXPECTED_ROUTE_SPAN_NAME)
-          .beginChildren()
-          .child(EXPECTED_LOGGER_SPAN_NAME)
-          .child(EXPECTED_RAISE_ERROR_SPAN)
-          .endChildren()
-          .child(EXPECTED_ROUTE_SPAN_NAME)
-          .beginChildren()
-          .child(EXPECTED_LOGGER_SPAN_NAME)
-          .child(EXPECTED_RAISE_ERROR_SPAN)
+          .addAttributesToAssertValue(createAttributeMap("round-robin-flow/processors/0/route/0/processors/1", TEST_ARTIFACT_ID))
+          .addAttributesToAssertExistence(attributesToAssertExistence)
           .endChildren()
           .endChildren()
-          .child(EXPECTED_ON_ERROR_PROPAGATE_SPAN_NAME)
+          .child(EXPECTED_ON_ERROR_PROPAGATE_SPAN)
+          .addAttributesToAssertValue(createAttributeMap("unknown", TEST_ARTIFACT_ID))
+          .addAttributesToAssertExistence(attributesToAssertExistence)
           .endChildren();
 
       expectedSpanHierarchy.assertSpanTree();
+      exportedSpans.forEach(span -> assertThat(span.getServiceName(), equalTo(span.getAttributes().get(ARTIFACT_ID_KEY))));
     } finally {
       spanCapturer.dispose();
     }
