+++ tracing/src/test/java/org/mule/test/components/tracing/TryScopeErrorOpenTelemetryTracingTestCase.java
@@ -4,6 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
+
 package org.mule.test.components.tracing;
 
 import static org.mule.test.allure.AllureConstants.Profiling.PROFILING;
@@ -13,7 +14,6 @@ import static org.mule.test.infrastructure.profiling.tracing.TracingTestUtils.cr
 import static org.mule.test.infrastructure.profiling.tracing.TracingTestUtils.getDefaultAttributesToAssertExistence;
 
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.hasSize;
 import static org.junit.Assert.assertThat;
 
 import org.mule.functional.junit4.MuleArtifactFunctionalTestCase;
@@ -36,36 +36,36 @@ import org.junit.Test;
 
 @Feature(PROFILING)
 @Story(DEFAULT_CORE_EVENT_TRACER)
-public class RoundRobinErrorTracingTestCase extends MuleArtifactFunctionalTestCase implements TracingTestRunnerConfigAnnotation {
+public class TryScopeErrorOpenTelemetryTracingTestCase extends MuleArtifactFunctionalTestCase
+    implements OpenTelemetryTracingTestRunnerConfigAnnotation {
 
   private static final int TIMEOUT_MILLIS = 30000;
   private static final int POLL_DELAY_MILLIS = 100;
 
-  public static final String EXPECTED_ROUTE_SPAN_NAME = "mule:round-robin:route";
-  public static final String EXPECTED_ROUND_ROBIN_SPAN_NAME = "mule:round-robin";
   public static final String EXPECTED_LOGGER_SPAN_NAME = "mule:logger";
-  public static final String ROUND_ROBIN_FLOW = "round-robin-flow";
+  public static final String TRY_SCOPE_FLOW = "try-scope-flow";
   public static final String EXPECTED_FLOW_SPAN_NAME = "mule:flow";
-  public static final String EXPECTED_RAISE_ERROR_SPAN = "mule:raise-error";
-  public static final String EXPECTED_ON_ERROR_PROPAGATE_SPAN = "mule:on-error-propagate";
+  public static final String EXPECTED_TRY_SCOPE_SPAN_NAME = "mule:try";
+  public static final String EXPECTED_RAISE_ERROR_SPAN_NAME = "mule:raise-error";
+  public static final String EXPECTED_SET_PAYLOAD_SPAN_NAME = "mule:set-payload";
+  public static final String EXPECTED_ON_ERROR_PROPAGATE_SPAN_NAME = "mule:on-error-propagate";
   public static final String NO_PARENT_SPAN = "0000000000000000";
-
-  public static final String TEST_ARTIFACT_ID = "RoundRobinErrorTracingTestCase#testRoundRobinFlowWithError";
+  public static final String TEST_ARTIFACT_ID = "TryScopeErrorOpenTelemetryTracingTestCase#testTryScope";
 
   @Inject
   PrivilegedProfilingService profilingService;
 
   @Override
   protected String getConfigFile() {
-    return "tracing/round-robin-error.xml";
+    return "tracing/try-scope-error.xml";
   }
 
   @Test
-  public void testRoundRobinFlowWithError() throws Exception {
+  public void testTryScope() throws Exception {
     ExportedSpanSniffer spanCapturer = profilingService.getSpanExportManager().getExportedSpanSniffer();
 
     try {
-      flowRunner(ROUND_ROBIN_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD).runExpectingException();
+      flowRunner(TRY_SCOPE_FLOW).withPayload(AbstractMuleTestCase.TEST_PAYLOAD).runExpectingException();
 
       PollingProber prober = new PollingProber(TIMEOUT_MILLIS, POLL_DELAY_MILLIS);
 
@@ -89,28 +89,22 @@ public class RoundRobinErrorTracingTestCase extends MuleArtifactFunctionalTestCa
 
       SpanTestHierarchy expectedSpanHierarchy = new SpanTestHierarchy(exportedSpans);
       expectedSpanHierarchy.withRoot(EXPECTED_FLOW_SPAN_NAME)
-          .addAttributesToAssertValue(createAttributeMap("round-robin-flow", TEST_ARTIFACT_ID))
+          .addAttributesToAssertValue(createAttributeMap("try-scope-flow", TEST_ARTIFACT_ID))
           .addAttributesToAssertExistence(attributesToAssertExistence)
           .beginChildren()
-          .child(EXPECTED_ROUND_ROBIN_SPAN_NAME)
-          .addAttributesToAssertValue(createAttributeMap("round-robin-flow/processors/0", TEST_ARTIFACT_ID))
-          .addAttributesToAssertExistence(attributesToAssertExistence)
-          .beginChildren()
-          .child(EXPECTED_ROUTE_SPAN_NAME)
-          .addAttributesToAssertValue(createAttributeMap("round-robin-flow/processors/0", TEST_ARTIFACT_ID))
+          .child(EXPECTED_TRY_SCOPE_SPAN_NAME)
+          .addAttributesToAssertValue(createAttributeMap("try-scope-flow/processors/0", TEST_ARTIFACT_ID))
           .addAttributesToAssertExistence(attributesToAssertExistence)
           .beginChildren()
           .child(EXPECTED_LOGGER_SPAN_NAME)
-          .addAttributesToAssertValue(createAttributeMap("round-robin-flow/processors/0/route/0/processors/0", TEST_ARTIFACT_ID))
+          .addAttributesToAssertValue(createAttributeMap("try-scope-flow/processors/0/processors/0", TEST_ARTIFACT_ID))
           .addAttributesToAssertExistence(attributesToAssertExistence)
-          .child(EXPECTED_RAISE_ERROR_SPAN)
-          .addAttributesToAssertValue(createAttributeMap("round-robin-flow/processors/0/route/0/processors/1", TEST_ARTIFACT_ID))
+          .child(EXPECTED_RAISE_ERROR_SPAN_NAME)
+          .addAttributesToAssertValue(createAttributeMap("try-scope-flow/processors/0/processors/1", TEST_ARTIFACT_ID))
           .addAttributesToAssertExistence(attributesToAssertExistence)
+          .child(EXPECTED_ON_ERROR_PROPAGATE_SPAN_NAME)
           .endChildren()
-          .endChildren()
-          .child(EXPECTED_ON_ERROR_PROPAGATE_SPAN)
-          .addAttributesToAssertValue(createAttributeMap("unknown", TEST_ARTIFACT_ID))
-          .addAttributesToAssertExistence(attributesToAssertExistence)
+          .child(EXPECTED_ON_ERROR_PROPAGATE_SPAN_NAME)
           .endChildren();
 
       expectedSpanHierarchy.assertSpanTree();
